Update README.md
Development of a Digital Chess Clock System 

Chess, a timeless strategy game, has long relied on analog chess clocks for timed play. As the game evolves in the digital age, there is a growing need for a modern, versatile, and user-friendly digital chess clock system. 
The aim of this mini project, "Development of a Digital Chess Clock System," is to address the shortcomings of traditional chess clocks and provide a solution that enhances the chess playing experience. With the advent of technology, we seek to offer a convenient and feature-rich alternative for chess enthusiasts and tournament organizers. 
This project introduces a state-of-the-art digital chess clock system that combines precision, usability, and flexibility. The key contributions include: 
User-Friendly Interface: Our system features an intuitive LCD interface with easily accessible control buttons, ensuring straightforward operation for players of all levels. 
Customizable Time Controls: Chess players can tailor time settings to suit their preferences, incorporating delay and increment options for added tactical depth. 
Dual-Clock Mode: The system supports competitive play with dual clocks, making it ideal for tournament use. 
Accessibility Features: To promote inclusivity, we have integrated audio cues, making the system accessible to visually impaired players. 
Win probability : The system predicts the winning probability based on the time gap between two players. Adding onto this a time difference is created for players who has lesser rating, so that weaker one gets   extra time to think. For this to decide, Individuals Rating is taken into account.Our project not only modernizes the traditional chess clock but also serves as a potential solution for other strategy-based board games. It offers a comprehensive and innovative tool that not only caters to the evolving needs of the chess community but also paves the way for the integration of digital technology into traditional board gaming